agents/langchain-chat-completions/pyproject.toml
[project]
name = "beamlit_simple_agent"
version = "0.1.0"
description = "Beamlit langchain agents"
authors = [
    { name = "cploujoux", email = "cploujoux@beamlit.com" },
]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "langchain-community==0.3.3",
    "langchain-openai==0.2.4",
    "langchain-mistralai==0.2.0",
    "langchain-core==0.3.13",
    "langgraph==0.2.40",
]

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
